fix: extract shared fields to BaseUserRequest and remove duplicate resolveMunicipalityId to resolve SonarCloud duplication warnings
