+++ src/main/java/se/sundsvall/users/api/model/UpdateUserRequest.java
@@ -4,36 +4,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import java.util.Objects;
-import se.sundsvall.dept44.common.validators.annotation.ValidMobileNumber;
-import se.sundsvall.users.api.validation.ValidEnum;
-import se.sundsvall.users.api.validation.ValidMunicipalityName;
-import se.sundsvall.users.integration.db.model.enums.Role;
-import se.sundsvall.users.integration.db.model.enums.Status;
 
-public class UpdateUserRequest {
+public class UpdateUserRequest extends BaseUserRequest {
 
 	@Schema(description = "E-postadress", example = "")
 	@NotBlank(message = "cannot be blank")
 	@Email
 	private String email;
-
-	@Schema(description = "Telefonnummer", example = "0701740669")
-	@NotBlank(message = "cannot be blank")
-	@ValidMobileNumber(message = "must be a valid mobile number")
-	private String phoneNumber;
-
-	@Schema(description = "Kommunnamn", example = "Sundsvall")
-	@NotBlank(message = "cannot be blank")
-	@ValidMunicipalityName(message = "must be a valid municipality name")
-	private String municipalityName;
-
-	@Schema(description = "Status", example = "ACTIVE")
-	@ValidEnum(message = "must be ACTIVE, INACTIVE or SUSPENDED", enumClass = Status.class, ignoreCase = true)
-	private String status;
-
-	@Schema(description = "Roll", example = "USER")
-	@ValidEnum(message = "must be USER or ADMIN", enumClass = Role.class, ignoreCase = true)
-	private String role;
 
 	public static UpdateUserRequest create() {
 		return new UpdateUserRequest();
@@ -52,61 +29,29 @@ public class UpdateUserRequest {
 		return this;
 	}
 
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
 	public UpdateUserRequest withPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
+		setPhoneNumber(phoneNumber);
 		return this;
-	}
-
-	public String getMunicipalityName() {
-		return municipalityName;
-	}
-
-	public void setMunicipalityName(String municipalityName) {
-		this.municipalityName = municipalityName;
 	}
 
 	public UpdateUserRequest withMunicipalityName(String municipalityName) {
-		this.municipalityName = municipalityName;
+		setMunicipalityName(municipalityName);
 		return this;
-	}
-
-	public String getStatus() {
-		return status;
-	}
-
-	public void setStatus(String status) {
-		this.status = status;
 	}
 
 	public UpdateUserRequest withStatus(String status) {
-		this.status = status;
+		setStatus(status);
 		return this;
 	}
 
-	public String getRole() {
-		return role;
-	}
-
-	public void setRole(String role) {
-		this.role = role;
-	}
-
 	public UpdateUserRequest withRole(String role) {
-		this.role = role;
+		setRole(role);
 		return this;
 	}
 
 	@Override
 	public int hashCode() {
-		return Objects.hash(email, municipalityName, phoneNumber, status, role);
+		return Objects.hash(super.hashCode(), email);
 	}
 
 	@Override
@@ -118,10 +63,6 @@ public class UpdateUserRequest {
 			return false;
 		}
 		UpdateUserRequest that = (UpdateUserRequest) o;
-		return Objects.equals(email, that.email)
-			&& Objects.equals(phoneNumber, that.phoneNumber)
-			&& Objects.equals(status, that.status)
-			&& Objects.equals(municipalityName, that.municipalityName)
-			&& Objects.equals(role, that.role);
+		return super.equals(o) && Objects.equals(email, that.email);
 	}
 }
